DjangoAPI/requirements.txt
APScheduler==3.10.0
asgiref==3.6.0
backports.zoneinfo==0.2.1
certifi==2022.12.7
charset-normalizer==3.0.1
click==8.1.3
colorama==0.4.6
Django==4.1.6
django-apscheduler==0.6.2
django-cors-headers==3.13.0
djangorestframework==3.14.0
h11==0.14.0
idna==3.4
psycopg2==2.9
pytz==2022.7.1
pytz-deprecation-shim==0.1.0.post0
requests==2.28.2
six==1.16.0
sqlparse==0.4.3
tzdata==2022.7
tzlocal==4.2
urllib3==1.26.14
uvicorn==0.20.0
